fix(slurmd): increase max cyclomatic complexity to 15
Ignoring an empty 'Nodes=' field in the Partition configuration by checking
the length of configured `partition_nodes` to be greater than 0, the cyclomatic
complexity of `interface_slurmd.Slurmd.get_new_nodes_and_nodes_and_partitions`
becomes 11 which cause the lint job to fail.

This method should eventually be reduced if possible, but setting the
max cyclomatic complexity to 15 works as an effective stop-gap.

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>