fix(slurmctld,slurmdbd): correct type hints for pyright
`pyright` doesn't like the dynamic attributes provided by the `slurmutils`
config models in the `slurmdbd` charm, so we'll need to enable type checking
over on that project.

For the `slurmctld` charm, we just needed to correct the hint for
`assemble_slurmctld_parameters` to be `dict[str, Any]` rather than
string which is returned for the slurm.conf editor.

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>