fix(ci): do not use destructive packing when building charms for release
Sets the `destructive-mode` option for the `upload-charm` action to false.
Pulling the down the base LXD instance will add some extra time onto the
workflow run, but it make us not be dependent on the base OS of the
GitHub runner VM.

Eventually we'll either want to replace the `cryptography` dependency
that takes so long to build, or use the `ccc` tool to cache the Slurm
charms dependencies so that the build for them is quicker.

Signed-off-by: Jason C. Nucciarone <nuccitheboss@ubuntu.com>